% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCytoscapeVizMap.R
\name{cytoscapeExport}
\alias{cytoscapeExport}
\title{Prepare networks for visualization in Cytoscape}
\usage{
cytoscapeExport(networkList, outputDir)
}
\arguments{
\item{networkList}{A named list of networks generated by the 
\code{\link[SetRank]{setRankAnalysis}} function.}

\item{outputDir}{The path where the files should be written to. The 
last element in the path will be created when it doesn't exist.}
}
\value{
None. The function writes out files as a side effect.
}
\description{
Exports a list of networks in GML format and creates a Cytoscape a VizMap 
XML file with a visualisation style applicable to all networks.
}
\details{
Each network will be written to a separate file in the specified output 
directory. The names of the supplied list of networks will be used as
filenames with the \code{.gml} extension added. The VizMap file will be 
called \code{setrank.xml} and will also be written to the output directory.
}
\author{
Cedric Simillion
}

