% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_nets.R
\name{fixGraphML}
\alias{fixGraphML}
\title{Fix the igraph graphML export}
\usage{
fixGraphML(fileName)
}
\description{
This function only exists to fix some issues with the graphML output of the 
igraph module.
}

