% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCytoscapeVizMap.R
\name{generateContinuousMapping}
\alias{generateContinuousMapping}
\title{Generates a continuous VizMap property mapping.}
\usage{
generateContinuousMapping(attributeName, visualValues, attributeValues,
  attributeType = "float", lesserValue = NA, greaterValue = NA)
}
\arguments{
\item{attributeName}{The node or edge attribute for which to create a visual
property mapping.}

\item{visualValues}{The values the visual property should have a key points.}

\item{attributeValues}{The key point attribute values}

\item{attributeType}{The type of the attribute. Must be either "\code{float}" or 
"\code{integer}".}

\item{lesserValue}{Optional. The value for the visual property for attribute
values below the lowest key point specified in \code{attributeValues}.}

\item{greaterValue}{Optional. The value for the visual property for attribute
values above the highest key point specified in \code{attributeValues}.}
}
\value{
An xmlNode object representing the VizMap property mapping.
}
\description{
Used by the \code{createCytoscapeVizMap} function. For internal use only.
}

