% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{membershipTable}
\alias{membershipTable}
\title{Create a gene set membership table.}
\usage{
membershipTable(selectedGenes, collection, network)
}
\arguments{
\item{selectedGenes}{The set of significant genes used during the SetRank
analysis.}

\item{collection}{The setCollection used during the SetRank analysis.}

\item{network}{A network generated by the 
\code{\link[SetRank]{setRankAnalysis}} function.}
}
\description{
Creates a table showing which significant genes belong to which significant
gene sets. This table allows to investigate the results of a SetRank 
analysis in more detail.
}
\author{
Cedric Simillion
}
\references{
A matrix of boolean values. The rownames are the geneIDs, the
column names are the gene set IDs. A \code{TRUE} value in a cell indicates
that the gene of the row is present in the gene set of the column; 
\code{FALSE} means otherwise.
}

