% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCytoscapeVizMap.R
\docType{data}
\name{propertyFunctions}
\alias{propertyFunctions}
\title{Attribute handlers for generating the VizMap XML}
\description{
A list of functions. Each name corresponds to a node or edge visual
property. The function tells the \code{\link{createCytoscapeVizMap}} 
function what type of attribute mapping to generate.
}

